Move the memory trace operations to their own module
They are all exported, so make it clear that they can be used in many places
