+++ pate.cabal
@@ -103,6 +103,7 @@ library
                        Pate.MemCell,
                        Pate.Memory,
                        Pate.Memory.Trace.MemTrace,
+                       Pate.Memory.Trace.Operations,
                        Pate.Memory.MemTrace,
                        Pate.Metrics,
                        Pate.SimulatorRegisters,
